Fix: an error was thrown if there was no config file before an release update

+++ homeControl.js
@@ -119,19 +119,23 @@ app.post('/updateApp', (req, res) => {
     fs.removeSync('./build.zip');
     fs.removeSync('./temp');
     download(zipURL, './build.zip', (path) => {
+        var backedupConfig = false;
         var zip = new AdmZip(path);
         zip.extractAllTo('./temp');
         try {
             fs.accessSync(path, fs.F_OK);
             fs.copySync('./deploy/config.js', '../config.js.back');
+            backedupConfig = true;
         }
         catch (e) {
             // It isn't accessible
         }
         fs.emptyDirSync('./deploy');
         fs.copySync('./temp', './deploy');
-        fs.copySync('../config.js.back', './deploy/config.js');
-        fs.remove('../config.js.back');
+        if (backedupConfig) {
+            fs.copySync('../config.js.back', './deploy/config.js');
+            fs.remove('../config.js.back');
+        }
         fs.removeSync('./build.zip');
         fs.removeSync('./temp');
         res.send('success');
